+++ src/components/Common/DomitoryManager/Modal/DataViewModal/index.tsx
