+++ src/components/Common/DomitoryManager/Modal/DataViewModal/style.ts
