+++ src/components/Common/DomitoryManager/Modal/RejectModal/index.tsx
