+++ src/components/Common/DomitoryManager/Modal/RejectModal/style.ts
